Delete Engine if it already exists

+++ quickstartapp/run.sh
@@ -21,7 +21,10 @@ echo "Step 1: Passed"
 
 echo "Step 2. Create a new Engine from an Engine Template"
 
-rm -r MyRecommendation
+if [ -d MyRecommendation ]; then
+  rm -r MyRecommendation
+fi
+
 git clone --depth 1 https://github.com/apache/incubator-predictionio-template-recommender.git MyRecommendation
 cd MyRecommendation
